+++ src/main/java/com/itechart/stockOnline/model/StockOwnerCompany.java
@@ -11,7 +11,7 @@ public class StockOwnerCompany {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name")
 //    @Size(max=50, message="Number of letters in name < 50")
@@ -45,8 +45,14 @@ public class StockOwnerCompany {
         admin = new User();
         this.name = name;
     }
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
